feat: implement JWT authentication, refactor API, and improve code structure
- Added `authController` for handling login and password generation functionality.
- Moved `CardProps` and `UserProps` types to dedicated files in `shared`.
- Implemented `JwtPayloadProps` for JWT payload handling.
- Secured routes with `authenticateJwt` middleware.
- Refactored game and authentication logic into separate controllers (`gameController`, `authController`).
- Moved JWT secret to a central configuration file (`config/index.ts`).
- Updated client-side API to handle JWT tokens in requests.
- Cleaned up and improved the structure of existing code and types.
