+++ src/client/components/handlers/CardRows/CardRows.tsx
@@ -2,7 +2,7 @@ import NoResultsMessage from '../../typography/texts/NoResultsMessage'
 import Button from '../../elements/Button/Button'
 import RouterLink from '../../typography/links/RouterLink'
 import TableContentContainer from '../../elements/Container/TableContentContainer'
-import { CardProps } from '../../../../shared/types'
+import { CardProps } from '../../../../shared/CardProps'
 import CardItem from '../../elements/CardItem/CardItem'
 
 type CardRowsProps = {
