+++ src/config/index.ts
@@ -1,0 +1,4 @@
+// TODO(fjv): Temporäre Lösung für eine Zentrale Konfigurationsdatei. Später in einer .env (dotenv) Datei?
+
+// TODO(fjv): Sollte das Secret eventuell länger und komplexer sein?
+export const JWT_SECRET = 'MyVerySecureSecret'
