+++ src/shared/CardProps.ts
@@ -1,0 +1,5 @@
+export type CardProps = {
+  id: string
+  front: string
+  back: string
+}
