+++ src/shared/JwtPayloadProps.ts
@@ -1,0 +1,4 @@
+export type JwtPayloadProps = {
+  username: string
+  role: string
+}
