+++ src/shared/UserProps.ts
@@ -1,9 +1,3 @@
-export type CardProps = {
-  id: string
-  front: string
-  back: string
-}
-
 export type UserProps = {
   username: string
   role: string
